# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is API for system voucher.",
        "title": "Sistem Voucher",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://academy.lumoshive.com/contact-us",
            "email": "lumoshive.academy@gmail.com"
        },
        "license": {
            "name": "Lumoshive Academy",
            "url": "https://academy.lumoshive.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/vouchers": {
            "get": {
                "security": [
                    {
                        "token": []
                    },
                    {
                        "id_key": []
                    }
                ],
                "description": "List Get Voucher By Query Param.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get Voucher By Query Param"
                ],
                "summary": "Get Voucher By Query Param",
                "responses": {
                    "200": {
                        "description": "successfully retrieved Voucher",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    },
                    "404": {
                        "description": "Voucher not found",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "token": []
                    },
                    {
                        "id_key": []
                    }
                ],
                "description": "Feature Use Voucher",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Use Voucher"
                ],
                "summary": "Use Voucher",
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request error",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/vouchers/:id": {
            "put": {
                "security": [
                    {
                        "token": []
                    },
                    {
                        "id_key": []
                    }
                ],
                "description": "Feature to update a voucher",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Update Voucher"
                ],
                "summary": "Update Voucher",
                "responses": {
                    "200": {
                        "description": "Success response with voucher data",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request error",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "token": []
                    },
                    {
                        "id_key": []
                    }
                ],
                "description": "Feature to delete a voucher",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Soft Deletes"
                ],
                "summary": "Soft Deletes",
                "responses": {
                    "200": {
                        "description": "Success response with voucher data",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request error",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/vouchers/:user_id": {
            "get": {
                "security": [
                    {
                        "token": []
                    },
                    {
                        "id_key": []
                    }
                ],
                "description": "Feature Find Voucher",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Find Voucher"
                ],
                "summary": "Find Voucher",
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request error",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    }
                }
            }
        },
...
        "/vouchers/create": {
            "post": {
                "security": [
                    {
                        "token": []
                    },
                    {
                        "id_key": []
                    }
                ],
                "description": "Feature to create a voucher",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Create Voucher"
                ],
                "summary": "Create Voucher",
                "responses": {
                    "200": {
                        "description": "Success response with voucher data",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request error",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    }
                }
            }
        },
        "/vouchers/redeem": {
            "post": {
                "security": [
                    {
                        "token": []
                    },
                    {
                        "id_key": []
                    }
                ],
                "description": "Feature to create a voucher",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Create Voucher"
                ],
                "summary": "Create Voucher",
                "responses": {
                    "200": {
                        "description": "Success response with voucher data",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request error",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    }
                }
            }
        },
...
        "/vouchers/redeem-points": {
            "get": {
                "security": [
                    {
                        "token": []
                    },
                    {
                        "id_key": []
                    }
                ],
                "description": "List Show Redeem Points.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Show Redeem Points"
                ],
                "summary": "Show Redeem Points",
                "responses": {
                    "200": {
                        "description": "successfully retrieved Voucher",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    },
                    "404": {
                        "description": "Voucher not found",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "helper.HTTPResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error_msg": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "id_key": {
            "type": "apiKey",
            "name": "id_key",
            "in": "header"
        },
        "token": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}